+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-13 14:03:21
-// Commit: refactor(core): optimize data flow between domain and data layers
+// Updated at: 2026-05-14 13:59:20
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '1103';
-  final String lastMessage = 'refactor(core): optimize data flow between domain and data layers';
-  final DateTime timestamp = DateTime.parse('2026-05-13T14:03:21+05:00');
+  final String commitRef = '1140';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-05-14T13:59:20+05:00');
 }